feat: update structure and add hierarchy services config
