+++ src/types/dashboard.d.ts
@@ -39,6 +39,7 @@ export interface LayoutConfig {
   children?: { name: string; children: LayoutConfig[]; expression?: string; enable?: boolean }[];
   activedTabIndex?: number;
   metricConfig?: MetricConfigOpt[];
+  hierarchyServicesConfig?: HierarchyServicesConfig[];
   id?: string;
   associate?: { widgetId: string }[];
   eventAssociate?: boolean;
@@ -81,6 +82,12 @@ export type MetricConfigOpt = {
   index?: number;
   detailLabel?: string;
 };
+
+export interface HierarchyServicesConfig {
+  layer: string;
+  nodeExpressions: string[];
+  expressionsConfig: MetricConfigOpt[];
+}
 
 export interface WidgetConfig {
   name?: string;
