+++ src/types/topology.d.ts
@@ -14,6 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 export interface Call {
   source: string | any;
   target: string | any;
