--- src/views/dashboard/related/topology/components/HierachySettings.vue
@@ -1,38 +1,0 @@
-<!-- Licensed to the Apache Software Foundation (ASF) under one or more
-contributor license agreements.  See the NOTICE file distributed with
-this work for additional information regarding copyright ownership.
-The ASF licenses this file to You under the Apache License, Version 2.0
-(the "License"); you may not use this file except in compliance with
-the License.  You may obtain a copy of the License at
-
-  http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License. -->
-<template>
-  <div class="hierarchy-services-settings">
-    <h5 class="title">{{ t("hierarchyServicesSettings") }}</h5>
-    <div class="label">{{ t("nodeMetrics") }}</div>
-  </div>
-</template>
-<script lang="ts" setup>
-  import { reactive, ref } from "vue";
-  import { useI18n } from "vue-i18n";
-
-  const { t } = useI18n();
-</script>
-<style lang="scss" scoped>
-  .title {
-    color: var(--sw-topology-color);
-    margin-bottom: 0;
-  }
-
-  .label {
-    font-size: $font-size-smaller;
-    margin-top: 10px;
-    color: var(--sw-topology-color);
-  }
-</style>
