+++ src/views/dashboard/related/topology/service/utils/layout.ts
